Introduce mediators.
The architecture is now as follows:

Artemis uses mediators to talk to devices. A mediator refers to a server
that Artemis uses to communicate with the devices. All devices must
eventually contact that server. This can be an MQTT server, or
Supabase (or anything else).

The mediators clients (which we simply call "mediators") are alive for the
duration of the lifetime of the Artemis instance. This means that Artemis
may do multiple requests on the same connection.