Improve authentication. (#148)
Instead of refreshing the token automatically when a supabase client is
started one now has to call `ensure_authenticated`.

When the authentication fails, remove the stored credentials.

Move the current-user from the http-server to the client.

After authentication store the current user-id of the http client in the
configuration. This makes it possible to run (most) tests for both the
http service as well as the supabase one.