Update goal fixes.
Use the term "goal" instead of "config"

The term "config" is overused, so we now use "goal" or "goal state" for
the entry on the broker that holds the requested configuration for a
device. That is, the "goal state" on the broker is the state a device
should try to get to.