Don't initialize the MQTT goal (#446)
There is little benefit in pushing an arbitrary goal state. On the
contrary; we would end up in a race condition where the initial (dummy)
goal state was pushed by the CLI, followed by a "real" goal state.
Depending on whether the service was fast enough in reading messages
from the broker it would either see the dummy goal state or not, making
the broker test flaky.