Remove unnecessary check.
We were testing that anon can download a file if it uses the public URL,
   but not if it uses the regular download.

It looks like newer versions of Supabase now also allow to download it
regularly, and tests broke. Since the data was already publicly
accessible there isn't a security issue and we can just drop the check.